Moved activitites to Ui folder. Added check in main activity for bluetooth enabled. Needs refactor
