--- app/src/main/java/com/android/ubclaunchpad/driver/MainActivity.java
@@ -1,13 +1,0 @@
-package com.android.ubclaunchpad.driver;
-
-import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-
-public class MainActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-    }
-}
