+++ app/src/main/java/com/android/ubclaunchpad/driver/UI/DispatchActivity.java
@@ -1,10 +1,9 @@
-package com.android.ubclaunchpad.driver;
+package com.android.ubclaunchpad.driver.UI;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 
-import com.android.ubclaunchpad.driver.login.LoginActivity;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 
@@ -21,13 +20,14 @@ public class DispatchActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
 
         FirebaseUser user = FirebaseAuth.getInstance().getCurrentUser();
-        if (user != null) {
+        //TODO revert if case before merge
+//        if (user != null) {
             // User is signed in
             startActivity(new Intent(this, MainActivity.class));
-        } else {
+//        } else {
             // No user is signed in
-            startActivity(new Intent(this, LoginActivity.class));
-        }
+//            startActivity(new Intent(this, LoginActivity.class));
+//        }
         finish();
     }
 }
