+++ app/src/main/java/com/android/ubclaunchpad/driver/UI/MainActivity.java
@@ -1,0 +1,49 @@
+package com.android.ubclaunchpad.driver.UI;
+
+import android.bluetooth.BluetoothAdapter;
+import android.content.Intent;
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+
+import com.android.ubclaunchpad.driver.R;
+
+public class MainActivity extends AppCompatActivity {
+
+    int REQUEST_ENABLE_BT = 1;
+
+    private boolean mBluetoothProblems = true;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+
+        bluetoothCheck(); //TODO does not really belong in view, move to model
+    }
+
+    @Override
+    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
+        if(requestCode == REQUEST_ENABLE_BT){
+            mBluetoothProblems = !(resultCode == RESULT_OK);
+        }
+    }
+
+    /**
+     * Checks if phone has bluetooth and if it is enabled
+     * If not enabled, prompt user to turn on
+     */
+    private void bluetoothCheck(){
+        //Enable Bluetooth
+        BluetoothAdapter mBluetoothAdapter = BluetoothAdapter.getDefaultAdapter();
+        if(mBluetoothAdapter == null){
+            //No bluetooth supported
+            mBluetoothProblems = true;
+            //TODO warn user they do not have bluetooth
+        }
+         else if (!mBluetoothAdapter.isEnabled()) {
+            // prompt user to turn on bluetooth
+            Intent enableBtIntent = new Intent(BluetoothAdapter.ACTION_REQUEST_ENABLE);
+            startActivityForResult(enableBtIntent, REQUEST_ENABLE_BT);
+        }
+    }
+}
