+++ app/src/main/java/com/android/ubclaunchpad/driver/login/LoginFragment.java
@@ -9,7 +9,7 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.android.ubclaunchpad.driver.MainActivity;
+import com.android.ubclaunchpad.driver.UI.MainActivity;
 import com.android.ubclaunchpad.driver.R;
 import com.facebook.login.LoginManager;
