+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.android.ubclaunchpad.driver.MainActivity">
+    tools:context=".UI.MainActivity">
 
     <TextView
         android:layout_width="wrap_content"
